move file position-n
