+++ __tests__/service.test.tsx
@@ -46,5 +46,4 @@ describe("Service", () => {
     });
     expect(servicesTitleContentButton).toBeInTheDocument();
   });
-
 });
