+++ styles/ServicesItem.module.css
@@ -8,7 +8,7 @@
   display: flex;
   flex-direction: column;
   width: 33%;
-  margin:63px 0;
+  margin: 63px 0;
 }
 
 .service_left_logo {
@@ -40,7 +40,7 @@
   width: 66%;
   flex-direction: column;
   border-left: 1px solid var(--border-default);
-  margin:63px 0;
+  margin: 63px 0;
   padding-left: 48px;
 }
