+++ styles/ServicesTitle.module.css
@@ -1,5 +1,5 @@
 .service_title {
-  padding:63px 0;
+  padding: 63px 0;
   border-bottom: 1px solid var(--border-default);
 }
 .service_title__heading {
